Reorganised STM project for the branch
This was a bad idea. I should've done develop first so everyone would have the updated project file.
