+++ UI_Operation.py
@@ -1,0 +1,136 @@
+import bpy
+import os
+from dataclasses import dataclass
+from .image_processing import outline_image
+
+@dataclass
+class PlaneItem:
+    PlaneFilepath = bpy.props.StringProperty(name="File Path",subtype='FILE_PATH')
+    PlaneRotation = bpy.props.IntProperty(name="Rotation", default=0)
+    ImagePlaneName: str
+    
+    def __init__(self, filepath ,rotation):
+        self.PlaneFilepath = filepath
+        self.PlaneRotation = rotation
+
+
+GlobalPlaneDataArray : list[PlaneItem] = [] #this will eventually replace the two array under this
+
+class Reset_Input_Images(bpy.types.Operator): 
+    bl_idname = "object.reset_selected_images"
+    bl_label = "Reset_Images"
+
+    def execute(self, context):
+        Itervalue = 0
+
+        for plane_data in GlobalPlaneDataArray :
+            #Finds the Images Saved
+            ImageFilePath = os.path.abspath("ImageFolder\\" + "View" + str(Itervalue) + plane_data.PlaneFilepath[plane_data.PlaneFilepath.rfind("."): ] ) 
+            # if we find that file we will delete it
+            if ImageFilePath: os.remove(ImageFilePath)
+            # selects the image plane in the array 
+            bpy.data.objects[plane_data.ImagePlaneName].select_set(True)
+            #deletes the image plane in the array
+            bpy.ops.object.delete(use_global=False, confirm=False)
+            #increases the itervale to reach the next View string
+            Itervalue = Itervalue + 1 
+
+        # clears the PlaneData Array
+        GlobalPlaneDataArray.clear() 
+        return {'FINISHED'}
+    
+
+    # Operator to add a new plane item
+class OBJECT_OT_add_plane_item(bpy.types.Operator):
+    bl_idname = "object.add_plane_item"
+    bl_label = "Add Plane Item"
+
+    def execute(self, context):
+        #adds the plane Itme to the Plane Item List
+        NewFileRotationPair = PlaneItem(bpy.context.scene.PlaneFilePath, bpy.context.scene.PlaneRotation )
+        GlobalPlaneDataArray.append(NewFileRotationPair)
+        return {'FINISHED'}
+
+    def draw(self, context):    
+        layout = self.layout   
+        row = layout.row()
+        row.prop(context.scene, "PlaneFilePath", text="FilePath : ") 
+        row = layout.row()
+        row.prop(context.scene, "PlaneRotation", text="Rotation : ", slider=True) 
+
+    def invoke(self, context, event):
+        return context.window_manager.invoke_props_dialog(self)
+
+
+class VIEW3D_PT_Sketch_To_Mesh_Views_FilePath_Panel(bpy.types.Panel):  
+    bl_label = "FilePath List"
+    bl_idname = "_PT_Views_File_Path"
+    bl_parent_id = "_PT_Views" 
+    bl_space_type = 'VIEW_3D'
+    bl_region_type = 'UI'
+
+    @classmethod
+    def poll(self,context):
+        return context.object is not None
+
+    def draw(self, context):
+        layout = self.layout
+        scene = context.scene
+
+        # List current plane items
+        for item in GlobalPlaneDataArray:
+            box = layout.box()
+            col = box.row()
+            filename = os.path.basename(item.PlaneFilepath)
+            col.label(text="Name: " + filename + "Rotation: " + str(item.PlaneRotation))
+
+        row = layout.row()
+        row.operator("object.reset_selected_images", text="Reset Images")
+
+
+class PlaceImageIn3D(bpy.types.Operator):
+    bl_idname = "object.place_image_in_space"
+    bl_label ="Place Images"
+
+    def execute(self, context):
+        #this will keep count of the views were have captured
+        Itervalue = 0
+        #this will be a folder in the sketch-to-Mesh project. This will hold the Image processed
+        ImageDiretoryForNewImage = "ImageFolder"
+        #this will eventually need to move to somewhere more accessable
+        #this is only here for now
+    
+        for plane_data in GlobalPlaneDataArray : 
+            if plane_data :
+                #this is used for the new image. We want to save the new image as the same type of file as the first
+                Extension =  plane_data.PlaneFilepath[plane_data.PlaneFilepath.rfind("."): ] 
+                #this is the file name for the image we are creating
+                plane_data.ImagePlaneName = "View" + str(Itervalue)
+                # allows us to access the plane after creation
+                #this is where we want to save the picture so i can be reaccessed
+                outline_image(plane_data.PlaneFilepath, Extension, plane_data.ImagePlaneName, ImageDiretoryForNewImage)
+                #this creates a new file path to the image we just saved
+                NewFilePath = os.path.abspath(ImageDiretoryForNewImage + "\\" + plane_data.ImagePlaneName + Extension) 
+                
+                if NewFilePath:
+                    filename = os.path.basename(NewFilePath)
+                    FileDirectory = NewFilePath[: NewFilePath.rfind("\\")] + "\\"
+                    #bpy.ops.import_image.to_plane(files=[{"name":filename, "name":filename}], directory=FileDirectory, relative=False)
+                    bpy.ops.import_image.to_plane(files=[{"name":filename, "name":filename}], directory=FileDirectory, relative=False)
+                    #we set the rotation and location of each plane
+                    bpy.data.objects[plane_data.ImagePlaneName].select_set(True)
+                    match Itervalue :
+                        case 1: bpy.ops.transform.translate(value=(-0.01, 0 , 0), orient_type='GLOBAL', orient_matrix=((1, 0, 0), (0, 1, 0), (0, 0, 1)), orient_matrix_type='GLOBAL', constraint_axis=(False, True, False), mirror=False, use_proportional_edit=False, proportional_edit_falloff='SMOOTH', proportional_size=1, use_proportional_connected=False, use_proportional_projected=False, snap=False, snap_elements={'INCREMENT'}, use_snap_project=False, snap_target='CLOSEST', use_snap_self=True, use_snap_edit=True, use_snap_nonedit=True, use_snap_selectable=False, alt_navigation=True)
+                        case 2: bpy.context.object.rotation_euler[2] = 0
+                else:
+                    match Itervalue:
+                        case 0: MissingView = "FontView"
+                        case 1: MissingView = "BackView"
+                        case 2: MissingView = "SideView"
+                    self.report({'ERROR'}, "No inputted Image for" + MissingView)
+                Itervalue = Itervalue + 1
+            else:
+                self.report({'ERROR'}, "No inputted Image.")
+                Itervalue = Itervalue + 1
+
+        return {'FINISHED'}
